refactor: Ranking -> RankingView 로 이름 변경
